{"schema":"decodex/commit/1","summary":"Split Radar signal rendering","authority":"XY-1115"}
